requirements.txt
# FastAPI and HTTP server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Vector Database
chromadb==0.4.18

# LLM and RAG Framework
langchain==0.1.0
langchain-community==0.0.10
groq==0.4.1

# Embeddings
sentence-transformers==2.2.2

# Document Processing
python-frontmatter==1.0.0
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
typing-extensions==4.8.0